migrated hot bar to new UI componenets

+++ src/dev/korgi/math/Vector2.java
@@ -1,13 +1,14 @@
 package dev.korgi.math;
 
 public class Vector2 {
-    public double x, y;
+    public float x, y;
 
-    public Vector2(double x, double y) {
+    public Vector2(float x, float y) {
         this.x = x;
         this.y = y;
     }
 
     public Vector2() {
     }
+
 }
